backend/pyproject.toml
[project]
name = "ai-trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-experimental>=0.3.4",
    "langchain-google-community>=2.0.7",
    "langchain-groq>=0.3.4",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.5.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ai-trip-planner",
    "black>=23.0.0",
    "fastapi>=0.104.0",
    "flake8>=6.0.0",
    "httpx>=0.25.0",
    "ipykernel>=6.29.5",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-experimental>=0.0.50",
    "langchain-google-community>=0.0.5",
    "langchain-groq>=0.0.5",
    "langchain-openai>=0.0.5",
    "langchain-tavily>=0.0.5",
    "langgraph>=0.0.30",
    "pydantic>=2.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "streamlit>=1.28.0",
    "uvicorn[standard]>=0.24.0",
]

[tool.uv.sources]
ai-trip-planner = { workspace = true }
